Moved the battlefield displays and txt printer to the utils folder, the Battlefield displays can now be called globaly.
