+++ src/com/clashofcards/renderer/AttackPhase.java
@@ -4,17 +4,25 @@ import com.clashofcards.Ai;
 import com.clashofcards.Card;
 import com.clashofcards.Player;
 import com.apps.util.Prompter;
+import com.clashofcards.utils.BattleFieldDisplay;
+import com.clashofcards.utils.Helper;
+
+import java.util.Iterator;
 import java.util.List;
 import java.util.Scanner;
 
 
 public class AttackPhase {
+    BattleFieldDisplay displayer = new BattleFieldDisplay();
 
     public void playerAttackPhase(Player player, Ai enemy, List<Card> playerBattleField, List<Card> enemyBattleField) {
-        displayEnemyBattleField(enemyBattleField);
-        showPlayerBattlefield(playerBattleField);
-        showPlayerHand(player.getDeck());
-        playCard(player.getDeck(), playerBattleField);
+        displayer.updateBattleField(enemyBattleField,playerBattleField,player);
+
+        // ATTACK METHOD IF THEY HAVE CARD,
+
+        playCard(player.getDeck(), playerBattleField); // player plays a card
+
+        displayer.updateBattleField(enemyBattleField,playerBattleField,player);
 
         boolean valid = false;
         while (!valid) {
@@ -34,50 +42,24 @@ public class AttackPhase {
                 System.out.println();
                 Prompter prompter = new Prompter(new Scanner(System.in));
                 String cardToPlay = prompter.prompt("Pick a card to play from your hand or pass (Enter the ID): ");
-                for (Card card : playerCards) {
+
+                // Use an iterator to avoid ConcurrentModificationException
+                Iterator<Card> iterator = playerCards.iterator();
+                while (iterator.hasNext()) {
+                    Card card = iterator.next();
                     if (card.getIndex().equals(Integer.valueOf(cardToPlay))) {
-                        playerCards.remove(card);
+                        iterator.remove(); // Remove the card from playerCards
                         playerBattlefield.add(card);
+                        System.out.println("You played: " + card.getName());
+                        Helper.delayGame(2);
                         valid = true;
-                    } else {
-                        System.out.println("Invalid card ID");
                     }
+                }
+
+                if (!valid) {
+                    System.out.println("Invalid card ID");
                 }
             }
         }
-    }
-
-    private void displayEnemyBattleField(List<Card> enemyBattleField) {
-        System.out.println();
-        if (enemyBattleField.isEmpty()) {
-            PrintBattlefieldText.printEnemyBattlefield();
-            System.out.println("----Enemy Has not played any cards yet----");
-        } else {
-            PrintBattlefieldText.printEnemyBattlefield();
-            for (Card card : enemyBattleField) {
-                System.out.println(card + " ");
-            }
-        }
-        System.out.println("\n" + "\n" + "\n");
-    }
-
-    private void showPlayerBattlefield(List<Card> playerBattlefield) {
-        if (playerBattlefield.isEmpty()) {
-            System.out.println("---Your battlefield is empty---");
-            PrintBattlefieldText.printYourBattleField();
-        } else {
-            for (Card card : playerBattlefield) {
-                System.out.print(card + "   ");
-            }
-            PrintBattlefieldText.printYourBattleField();
-            System.out.println("\n"+"\n"+"\n"+"\n");
-        }
-    }
-
-    private void showPlayerHand(List<Card> playerHand) {
-        for (Card card : playerHand) {
-            System.out.print(card + "   ");
-        }
-        PrintBattlefieldText.printYourHand();
     }
 }
